requirements.txt
reportlab>=4.0.4
Pillow>=10.0.0
pdfplumber>=0.10.2
nltk>=3.8.1
textblob>=0.17.1
python-docx>=0.8.11
ebooklib>=0.18
pyinstaller>=6.3.0 